+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Shell completion
 - Updated to Rust 2021
 
 ## [1.0.0] - 2021-08-27